fix probe skeleton load, improve failure dump rendering
replace 100ms poll freeze trigger with KVM hardware watchpoint on
sch->exit_kind (DR0 write watchpoint, DR7=0xD0602). the vCPU that
writes exit_kind takes #DB → KVM_EXIT_DEBUG → instant freeze with
all sched_ext state still valid. gate on kind >= SCX_EXIT_ERROR to
suppress false triggers from clean shutdown transitions. EINTR
retry with 3-failure backoff for transient set_guest_debug errors.

rewrite early scanner from per-rq runnable_list walk (drained by
bypass before observation) to global scx_tasks LIST_HEAD walk.
gate on SCX_TASK_QUEUED flag to exclude sleeping tasks with stale
runnable_at timestamps. cursor entries filtered via SCX_TASK_CURSOR.

add BPF .bss snapshot of sched_state at tp_btf/sched_ext_exit:
5 sticky vars (aborting, bypass_depth, exit_kind, sched_kva,
watchdog_timeout) captured before the CAS latch publishes. host
fallback decoder in dump_state resolves the snapshot when the live
scx_root walk returns None (scheduler already torn down).

convert 17 global atomic .bss counters to cacheline-aligned per-CPU
array (Pattern A: __attribute__((aligned(128))), indexed by
bpf_get_smp_processor_id() & CPU_MASK). eliminates cross-CPU
cacheline bouncing on preempt_disable/enable tracepoints (~millions
of LOCK'd adds/sec on 96-core hosts). host sums across CPU slots
at dump time. add ProbeBssCounters to failure dump report with E2E
test asserting trigger_count > 0 after stall.

performance fixes: u64-chunked volatile reads in reader.rs (8x on
hot path), BTF parse cached per-session instead of per-function at
probe attach, O(1) HashMap lookups replacing O(N²) scans in probe
dump and symbol resolution, reusable scratch buffers for virtio-net
TX and virtio-console RX, arena page buffer reuse on freeze path,
consecutive-fail cap on BTF type ID probe, sleep(1ms) replacing
yield_now() busy-spin in vcpu wait_for_exit.

+++ src/monitor/debug_capture.rs
@@ -318,7 +318,7 @@ pub enum AffinityHint {
     /// `thread_siblings_list` entry. Popcount == 1 →
     /// [`Self::SingleCpu`]. Popcount == 2 but the CPUs are NOT
     /// siblings → [`Self::Exact`]. Partial sibling sets (popcount
-    /// > 1 but < `threads_per_core` on SMT-N>2 hosts) project to
+    /// `> 1` but `< threads_per_core` on SMT-N>2 hosts) project to
     /// [`Self::Exact`], not `SmtSiblingPair`.
     SmtSiblingPair {
         #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -715,7 +715,7 @@ fn project_sched_policy_hints(dump: &FailureDumpReport) -> Vec<SchedPolicyHint> 
                 // projected `SchedPolicyHint::Other { nice }` is
                 // spawnable by the reproducer generator without
                 // further sanitisation.
-                let raw = (task.static_prio as i32) - 120;
+                let raw = task.static_prio - 120;
                 let nice = raw.clamp(-20, 19);
                 SchedPolicyHint::Other { nice }
             }
